+++ src/parser/block/reply.test.ts
@@ -1,10 +1,10 @@
-import { mention } from './mention';
+import { reply } from './reply';
 import { some } from '../../combinator';
 import { inspect } from '../../debug.test';
 
-describe('Unit: parser/block/mention', () => {
-  describe('mention', () => {
-    const parser = (source: string) => some(mention)(source, {});
+describe('Unit: parser/block/reply', () => {
+  describe('reply', () => {
+    const parser = (source: string) => some(reply)(source, {});
 
     it('basic', () => {
       assert.deepStrictEqual(inspect(parser('>>1')), [['<p><span class="cite">&gt;<a href="?comment=1" class="anchor" data-depth="1">&gt;1</a></span></p>'], '']);